requirements.txt
alembic==1.13.1
blinker==1.7.0
cachetools==5.3.2
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
dask==2024.2.0
emoji==2.10.1
filelock==3.9.0
Flask==3.0.2
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
flask-paginate==2023.10.24
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.49.0
fsspec==2024.2.0
google-api-core==2.17.1
google-api-python-client==2.118.0
google-auth==2.28.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.62.0
googletrans==4.0.0rc1
greenlet==3.0.3
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.2.1
httpcore==0.9.1
httplib2==0.22.0
httpx==0.13.3
huggingface-hub==0.20.3
hyperframe==5.2.0
idna==2.10
imbalanced-learn==0.12.0
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
langdetect==1.0.9
locket==1.0.0
Mako==1.3.2
MarkupSafe==2.1.5
matplotlib==3.8.3
mlxtend==0.23.1
mpmath==1.3.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.3
packaging==23.2
pandas==2.2.0
partd==1.4.1
pillow==10.2.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyMySQL==1.1.0
pyotp==2.9.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.28.1
rfc3986==1.5.0
rsa==4.9
safetensors==0.4.2
Sastrawi==1.0.1
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.27
stanza==1.7.0
sympy==1.12
textblob==0.18.0.post0
threadpoolctl==3.3.0
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
torch==2.2.0
torchaudio==2.2.0
torchvision==0.17.0
tqdm==4.66.2
transformers==4.37.2
typing_extensions==4.9.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.13
Werkzeug==3.0.1
wordcloud==1.9.3
WTForms==3.1.2
zipp==3.17.0
